se agregaron muchas cosas

+++ proyectocountertexdefinitivo/Models/PerfilEmpleado.cs
@@ -6,7 +6,7 @@ namespace proyectocountertexdefinitivo.Models
     public class PerfilEmpleado
     {
         public int IdEmpleado { get; set; }
-        public int ProduccionDiaria { get; set; }
+        public decimal ProduccionDiaria { get; set; }
         public string TipoPrenda { get; set; }
         public string TipoOperacion { get; set; }
         public int CantidadOperacion { get; set; }
@@ -15,7 +15,7 @@ namespace proyectocountertexdefinitivo.Models
         public DateTime ControlHorarios { get; set; }
         public DateTime HoraEntrada { get; set; }
         public DateTime HoraSalida { get; set; }
-        public int MetaPorCorte { get; set; }
+        public decimal MetaPorCorte { get; set; }
         public string BotonAyuda { get; set; }
         public string Estadisticas { get; set; }
         public string Observaciones { get; set; }
